+++ TestLibXPUInfo/TestLibXPUInfo.vcxproj.filters
@@ -18,12 +18,12 @@
     <ClCompile Include="TestLibXPUInfo.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\LibXPUInfo_D3D12Utility.cpp">
+    <ClCompile Include="..\utility\LibXPUInfo_D3D12Utility.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\LibXPUInfo_D3D12Utility.h">
+    <ClInclude Include="..\utility\LibXPUInfo_D3D12Utility.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
